requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
cohere>=5.0.0
wikipedia>=1.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

pytest>=7.4.0
pytest-asyncio>=0.21.0